And make sure the types actually check
darcs-hash:456ca0951fb957e749198015355114f70db026a6

+++ FastIO.hs
@@ -278,7 +278,7 @@ printfPS :: P.FastString -> Int -> Int -> P.FastString
 printfPS fmt arg1 arg2 =
     unsafePerformIO $ P.generate lim $ \ptr ->
         P.unsafeUseAsCString fmt $ \c_fmt -> do
-            sz' <- c_printf2d ptr lim (castPtr c_fmt)
+            sz' <- c_printf2d ptr (fromIntegral lim) (castPtr c_fmt)
                         (fromIntegral arg1) (fromIntegral arg2)
             return (min lim (fromIntegral sz')) -- snprintf might truncate
     where
